Added left right strafing instead of A and D being rotational
WASD controls usually allow for A and D to be strafing mechanisms instead of rotation. Although now it's impossible to play without a mouse!

Perhaps adding in the addition of arrow keys that move the player forward and back along with left and right being rotate will allow non-mouse users to play?